Finish converting to compatible non-xml.
Add the disabling of single-path unit upgrading.
